+++ examples/tetris/model/game.rb
@@ -36,9 +36,12 @@ class Tetris
       PREVIEW_PLAYFIELD_WIDTH = 4
       PREVIEW_PLAYFIELD_HEIGHT = 2
       SCORE_MULTIPLIER = {1 => 40, 2 => 100, 3 => 300, 4 => 1200}
+      UP_ARROW_ACTIONS = %i[instant_down rotate_right rotate_left]
+      SPEEDS = %i[snail sloth turtle rabbit gorilla bear horse gazelle cheetah falcon]
+      SPEED_INITIAL_DELAYS = SPEEDS.each_with_index.inject({}) {|hash, speed_index_pair| hash.merge(speed_index_pair.first => 1.1 - speed_index_pair.last*(0.1)) }
       
       attr_reader :playfield_width, :playfield_height
-      attr_accessor :game_over, :paused, :preview_tetromino, :lines, :score, :level, :high_scores, :beeping, :added_high_score, :show_high_scores, :up_arrow_action, :show_preview_tetromino
+      attr_accessor :game_over, :paused, :preview_tetromino, :lines, :score, :level, :high_scores, :beeping, :added_high_score, :show_high_scores, :up_arrow_action, :show_preview_tetromino, :initial_delay
       alias game_over? game_over
       alias paused? paused
       alias beeping? beeping
@@ -46,6 +49,7 @@ class Tetris
       alias show_preview_tetromino? show_preview_tetromino
       
       def initialize(playfield_width = PLAYFIELD_WIDTH, playfield_height = PLAYFIELD_HEIGHT)
+        @initial_delay = SPEED_INITIAL_DELAYS[:snail]
         @playfield_width = playfield_width
         @playfield_height = playfield_height
         @high_scores = []
@@ -202,35 +206,31 @@ class Tetris
       end
       
       def delay
-        [1.1 - (level.to_i * 0.1), 0.001].max
+        [@initial_delay - (level.to_i * 0.1), 0.001].max
       end
       
       def beep
         @beeper&.call if beeping
       end
-      
-      def instant_down_on_up=(value)
-        self.up_arrow_action = :instant_down if value
+
+      SPEED_INITIAL_DELAYS.each do |speed, speed_initial_day|
+        define_method("speed_#{speed}=") do |is_true|
+          self.initial_delay = speed_initial_day if is_true
+        end
+        
+        define_method("speed_#{speed}") do
+          self.initial_delay == speed_initial_day
+        end
       end
       
-      def instant_down_on_up
-        self.up_arrow_action == :instant_down
-      end
-      
-      def rotate_right_on_up=(value)
-        self.up_arrow_action = :rotate_right if value
-      end
-      
-      def rotate_right_on_up
-        self.up_arrow_action == :rotate_right
-      end
-      
-      def rotate_left_on_up=(value)
-        self.up_arrow_action = :rotate_left if value
-      end
-      
-      def rotate_left_on_up
-        self.up_arrow_action == :rotate_left
+      UP_ARROW_ACTIONS.each do |up_arrow_action_symbol|
+        define_method("#{up_arrow_action_symbol}_on_up=") do |is_true|
+          self.up_arrow_action = up_arrow_action_symbol if is_true
+        end
+        
+        define_method("#{up_arrow_action_symbol}_on_up") do
+          self.up_arrow_action == up_arrow_action_symbol
+        end
       end
       
       def reset_tetrominoes